+++ src/surface/edt-surface.js
@@ -9,100 +9,9 @@ import { VolumeSurface } from "./volume.js";
 import Grid from "../geometry/grid.js";
 import {
     computeBoundingBox,
-    v3subScalar, v3addScalar, v3divideScalar, v3multiplyScalar,
-    v3floor, v3ceil, v3sub, v3negate
 } from "../math/vector-utils.js";
-import {
-    m4new, m4multiply, m4makeTranslation, m4makeScale, m4makeRotationY
-} from "../math/matrix-utils.js";
-import { degToRad } from "../math/math-utils.js";
-import { getRadiusDict } from "./surface-utils.js";
+import { getRadiusDict, getSurfaceGrid } from "./surface-utils.js";
 
-
-function getSurfaceGrid( min, max, maxRadius, scaleFactor, extraMargin ){
-
-    // need margin to avoid boundary/round off effects
-    var margin = ( 1 / scaleFactor ) * 3;
-    margin += maxRadius;
-
-    v3subScalar( min, min, extraMargin + margin );
-    v3addScalar( max, max, extraMargin + margin );
-
-    v3multiplyScalar( min, min, scaleFactor );
-    v3floor( min, min );
-    v3divideScalar( min, min, scaleFactor );
-
-    v3multiplyScalar( max, max, scaleFactor );
-    v3ceil( max, max );
-    v3divideScalar( max, max, scaleFactor );
-
-    var dim = new Float32Array( 3 );
-    v3sub( dim, max, min );
-    v3multiplyScalar( dim, dim, scaleFactor );
-    v3ceil( dim, dim );
-    v3addScalar( dim, dim, 1 );
-
-    var maxSize = Math.pow( 10, 6 ) * 256;
-    var tmpSize = dim[ 0 ] * dim[ 1 ] * dim[ 2 ] * 3;
-
-    if( maxSize <= tmpSize ){
-
-        scaleFactor *= Math.pow( maxSize / tmpSize, 1/3 );
-
-        v3multiplyScalar( min, min, scaleFactor );
-        v3floor( min, min );
-        v3divideScalar( min, min, scaleFactor );
-
-        v3multiplyScalar( max, max, scaleFactor );
-        v3ceil( max, max );
-        v3divideScalar( max, max, scaleFactor );
-
-        v3sub( dim, max, min );
-        v3multiplyScalar( dim, dim, scaleFactor );
-        v3ceil( dim, dim );
-        v3addScalar( dim, dim, 1 );
-
-    }
-
-    var tran = new Float32Array( min );
-    v3negate( tran, tran );
-
-    // coordinate transformation matrix
-    var matrix = m4new();
-    var mroty = m4new();
-    m4makeRotationY( mroty, degToRad( 90 ) );
-    m4multiply( matrix, matrix, mroty );
-
-    var mscale = m4new();
-    m4makeScale( mscale,
-        -1 / scaleFactor,
-        1 / scaleFactor,
-        1 / scaleFactor
-    );
-    m4multiply( matrix, matrix, mscale );
-
-    var mtrans = m4new();
-    m4makeTranslation( mtrans,
-        -scaleFactor * tran[2],
-        -scaleFactor * tran[1],
-        -scaleFactor * tran[0]
-    );
-    m4multiply( matrix, matrix, mtrans );
-
-    return {
-        dim: dim,
-        tran: tran,
-        matrix: matrix,
-        scaleFactor: scaleFactor
-    };
-
-}
-getSurfaceGrid.__deps = [
-    degToRad,
-    v3subScalar, v3addScalar, v3divideScalar, v3multiplyScalar,
-    v3floor, v3ceil, v3sub, v3negate,
-    m4new, m4multiply, m4makeTranslation, m4makeScale, m4makeRotationY
-];
