+++ src/surface/surface-utils.js
@@ -4,10 +4,16 @@
  * @private
  */
 
-
+import { degToRad } from "../math/math-utils.js";
 import {
-    v3sub, v3cross, v3fromArray, normalizeVector3array
+    m4new, m4multiply, m4makeTranslation, m4makeScale, m4makeRotationY
+} from "../math/matrix-utils.js";
+import {
+    v3addScalar, v3subScalar, v3divideScalar, v3multiplyScalar,
+    v3floor, v3ceil,  v3sub, v3negate,
+    v3cross, v3fromArray, normalizeVector3array
 } from "../math/vector-utils.js";
+
 
 
 function laplacianSmooth( verts, faces, numiter, inflate ){
@@ -338,8 +344,96 @@ function getRadiusDict( radiusList ){
 }
 
 
+function getSurfaceGrid( min, max, maxRadius, scaleFactor, extraMargin ){
+
+    // need margin to avoid boundary/round off effects
+    var margin = ( 1 / scaleFactor ) * 3;
+    margin += maxRadius;
+
+    v3subScalar( min, min, extraMargin + margin );
+    v3addScalar( max, max, extraMargin + margin );
+
+    v3multiplyScalar( min, min, scaleFactor );
+    v3floor( min, min );
+    v3divideScalar( min, min, scaleFactor );
+
+    v3multiplyScalar( max, max, scaleFactor );
+    v3ceil( max, max );
+    v3divideScalar( max, max, scaleFactor );
+
+    var dim = new Float32Array( 3 );
+    v3sub( dim, max, min );
+    v3multiplyScalar( dim, dim, scaleFactor );
+    v3ceil( dim, dim );
+    v3addScalar( dim, dim, 1 );
+
+    var maxSize = Math.pow( 10, 6 ) * 256;
+    var tmpSize = dim[ 0 ] * dim[ 1 ] * dim[ 2 ] * 3;
+
+    if( maxSize <= tmpSize ){
+
+        scaleFactor *= Math.pow( maxSize / tmpSize, 1/3 );
+
+        v3multiplyScalar( min, min, scaleFactor );
+        v3floor( min, min );
+        v3divideScalar( min, min, scaleFactor );
+
+        v3multiplyScalar( max, max, scaleFactor );
+        v3ceil( max, max );
+        v3divideScalar( max, max, scaleFactor );
+
+        v3sub( dim, max, min );
+        v3multiplyScalar( dim, dim, scaleFactor );
+        v3ceil( dim, dim );
+        v3addScalar( dim, dim, 1 );
+
+    }
+
+    var tran = new Float32Array( min );
+    v3negate( tran, tran );
+
+    // coordinate transformation matrix
+    var matrix = m4new();
+    var mroty = m4new();
+    m4makeRotationY( mroty, degToRad( 90 ) );
+    m4multiply( matrix, matrix, mroty );
+
+    var mscale = m4new();
+    m4makeScale( mscale,
+        -1 / scaleFactor,
+        1 / scaleFactor,
+        1 / scaleFactor
+    );
+    m4multiply( matrix, matrix, mscale );
+
+    var mtrans = m4new();
+    m4makeTranslation( mtrans,
+        -scaleFactor * tran[2],
+        -scaleFactor * tran[1],
+        -scaleFactor * tran[0]
+    );
+    m4multiply( matrix, matrix, mtrans );
+
+    return {
+        dim: dim,
+        tran: tran,
+        matrix: matrix,
+        scaleFactor: scaleFactor
+    };
+
+}
+getSurfaceGrid.__deps = [
+    degToRad,
+    v3subScalar, v3addScalar, v3divideScalar, v3multiplyScalar,
+    v3floor, v3ceil, v3sub, v3negate,
+    m4new, m4multiply, m4makeTranslation, m4makeScale, m4makeRotationY
+];
+
+
+
 export {
     laplacianSmooth,
     computeVertexNormals,
-    getRadiusDict
+    getRadiusDict,
+    getSurfaceGrid
 };
